:sparkles: fix: mirage-new-json
